Handle 1v1, 2v2, rotating and all regions

+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/data/dto/StatDetailDto.kt
@@ -11,7 +11,7 @@ data class StatDetailDto(
     @SerialName("name")
     val name: String,
     @SerialName("xp")
-    val xp: Double,
+    val xp: Int,
     @SerialName("level")
     val level: Int,
     @SerialName("xp_percentage")
